Restructuring into multiple files
