+++ main.go
@@ -5,14 +5,8 @@ import (
 	"io"
 	"log"
 	"os"
-	"path/filepath"
 	"runtime"
 	"strings"
-	"sync"
-	"time"
-
-	"github.com/idelchi/wslint/internal/linter"
-	"github.com/idelchi/wslint/matcher"
 )
 
 //nolint:gochecknoglobals // Global variable for CI stamping.
@@ -33,16 +27,6 @@ func exit(code int, msg string) {
 	}
 
 	os.Exit(code)
-}
-
-// Options contains the options for the linter.
-type Options struct {
-	Exclude         []string
-	NumberOfWorkers int
-	Fix             bool
-	Logger          *log.Logger
-	Patterns        []string
-	Hidden          bool
 }
 
 //nolint:funlen // This function is long, but it's not complex.
@@ -112,183 +96,4 @@ func main() {
 	}
 
 	os.Exit(match(options))
-}
-
-func match(options Options) int {
-	verboseLog := options.Logger
-	patterns := options.Patterns
-	hidden := options.Hidden
-	exclude := options.Exclude
-
-	// Create a matcher
-	matcher := matcher.New(hidden, exclude, verboseLog)
-
-	// Collect the files to inspect, ranging over the patterns
-	for _, arg := range patterns {
-		if err := matcher.Match(arg); err != nil {
-			log.Printf("Error: %v", err)
-
-			return 1
-		}
-	}
-
-	// Create a slice of files to inspect
-	files := []linter.Linter{}
-
-	// Fill the slice with files
-	for _, file := range matcher.ListFiles() {
-		// Get the relative path to the execution directory
-		if fileRel, err := filepath.Rel(".", file); err == nil {
-			file = fileRel
-		}
-
-		lint := linter.NewLinter(file)
-
-		// Append the linter to the slice
-		files = append(files, *lint)
-
-		verboseLog.Printf("<included> %q", file)
-	}
-
-	// If no files are found, exit with error code 0
-	if len(files) == 0 {
-		log.Println("No files found")
-
-		return 1
-	}
-
-	return process(options, files)
-}
-
-func process(options Options, files []linter.Linter) int {
-	numberOfFiles := len(files)
-	options.Logger.Printf("Processing %d files", numberOfFiles)
-
-	if numberOfFiles < options.NumberOfWorkers {
-		options.NumberOfWorkers = numberOfFiles
-	}
-
-	workerPool := WorkerPool{
-		NumberOfWorkers: options.NumberOfWorkers,
-		NumberOfJobs:    numberOfFiles,
-		Fix:             options.Fix,
-		Files:           files,
-		Logger:          options.Logger,
-	}
-
-	// Create channels for sending and receiving jobs and results
-	jobs := make(chan linter.Linter, numberOfFiles)
-	results := make(chan linter.Linter, numberOfFiles)
-
-	workerPool.Start(jobs, results)
-
-	exitCode := 0
-
-	// Collect the results
-	for range files {
-		result := <-results
-
-		if ok := result.Summary(); !ok {
-			exitCode = 1
-		}
-	}
-
-	if exitCode == 0 {
-		log.Println("No issues found")
-	}
-
-	return exitCode
-}
-
-// WorkerPool represents a pool of workers.
-type WorkerPool struct {
-	// The number of workers in the pool
-	NumberOfWorkers int
-	// The number of jobs to process
-	NumberOfJobs int
-	// Logger
-	Logger *log.Logger
-	// Fix
-	Fix bool
-	// Files
-	Files []linter.Linter
-}
-
-// Start the worker pool.
-func (p *WorkerPool) Start(jobs, results chan linter.Linter) {
-	// Create a wait group to ensure all workers have finished
-	var waitGroup sync.WaitGroup
-
-	// Start the workers
-	for i := 0; i < p.NumberOfWorkers; i++ {
-		waitGroup.Add(1)
-
-		i := i
-
-		go func() {
-			defer waitGroup.Done()
-			worker(i+1, p.Logger, p.Fix, jobs, results)
-		}()
-	}
-
-	// Measure the time it takes to process all the files
-	start := time.Now()
-
-	// Send the jobs to the workers
-	for _, file := range p.Files {
-		jobs <- file
-	}
-
-	close(jobs)
-
-	// Wait for all the workers to finish
-	waitGroup.Wait()
-
-	// Measure the time it takes to process all the files
-	// TODO(Idelchi): Move this log message to the appear last.
-	p.Logger.Printf("<processed> all (%d) files in %s", len(p.Files), time.Since(start))
-}
-
-// TODO(Idelchi): Rework concurrency, read
-// https://twin.sh/articles/39/go-concurrency-goroutines-worker-pools-and-throttling-made-simple
-
-// worker processes jobs.
-func worker(
-	identifier int,
-	logger *log.Logger,
-	fix bool,
-	files <-chan linter.Linter,
-	results chan<- linter.Linter,
-) {
-	jobsProcessed := 0
-
-	for file := range files {
-		logger.Printf("<processing> %q", file.Name)
-
-		func() {
-			var err error
-			defer func() {
-				file.Error = err
-			}()
-
-			reader := &linter.Reader{}
-
-			if file.Error = file.Lint(reader); file.Error != nil {
-				return
-			}
-
-			if fix && file.HasIssues() {
-				writer := &linter.Writer{
-					Main: reader,
-				}
-				err = file.Fix(writer)
-			}
-		}()
-
-		results <- file
-
-		jobsProcessed++
-	}
-
-	logger.Printf("<worker %d> processed %d jobs", identifier, jobsProcessed)
 }
