+++ worker.go
@@ -1,0 +1,101 @@
+package main
+
+import (
+	"log"
+	"sync"
+	"time"
+
+	"github.com/idelchi/wslint/internal/linter"
+)
+
+// WorkerPool represents a pool of workers.
+type WorkerPool struct {
+	// The number of workers in the pool
+	NumberOfWorkers int
+	// The number of jobs to process
+	NumberOfJobs int
+	// Logger
+	Logger *log.Logger
+	// Fix
+	Fix bool
+	// Files
+	Files []linter.Linter
+	// Time spent processing the files
+	ProcessingTime time.Duration
+}
+
+// Start the worker pool.
+func (p *WorkerPool) Start(jobs, results chan linter.Linter) {
+	// Create a wait group to ensure all workers have finished
+	var waitGroup sync.WaitGroup
+
+	// Start the workers
+	for i := 0; i < p.NumberOfWorkers; i++ {
+		waitGroup.Add(1)
+
+		i := i
+
+		go func() {
+			defer waitGroup.Done()
+			worker(i+1, p.Logger, p.Fix, jobs, results)
+		}()
+	}
+
+	// Measure the time it takes to process all the files
+	start := time.Now()
+
+	// Send the jobs to the workers
+	for _, file := range p.Files {
+		jobs <- file
+	}
+
+	close(jobs)
+
+	// Wait for all the workers to finish
+	waitGroup.Wait()
+
+	// Measure the time it takes to process all the files
+	p.ProcessingTime = time.Since(start)
+}
+
+// Stats prints the stats of the worker pool run.
+func (p *WorkerPool) Stats() {
+	p.Logger.Printf("<processed> %d files in %s", len(p.Files), p.ProcessingTime)
+}
+
+// worker processes jobs.
+// https://twin.sh/articles/39/go-concurrency-goroutines-worker-pools-and-throttling-made-simple
+func worker(
+	identifier int,
+	logger *log.Logger,
+	fix bool,
+	files <-chan linter.Linter,
+	results chan<- linter.Linter,
+) {
+	jobsProcessed := 0
+
+	for file := range files {
+		logger.Printf("<processing> %q", file.Name)
+
+		func() {
+			reader := &linter.Reader{}
+
+			if file.Error = file.Lint(reader); file.Error != nil {
+				return
+			}
+
+			if fix && file.HasIssues() {
+				writer := &linter.Writer{
+					Main: reader,
+				}
+				file.Error = file.Fix(writer)
+			}
+		}()
+
+		results <- file
+
+		jobsProcessed++
+	}
+
+	logger.Printf("<worker %d> processed %d jobs", identifier, jobsProcessed)
+}
